+++ R/sankey.R
@@ -2,14 +2,17 @@
 #'
 #' Create a HTML sankey plot for displaying admissions and retention data
 #'
-#' @param data Data-frame containing one row per observation (student).
-#' @param steps Character vector or list containing a subset of the column names in `data`.
+#' @param   data   Data-frame containing one row per observation (student).
+#' @param   steps   Character vector or list containing a subset of the column names in `data`.
 #' These correspond to the stages of the sankey chart. Where a list is provided, elements of the
 #' list can contain vectors which specify how a step will be further subdivided (drilled-down into)
 #' when that step is clicked on.
-#' @param width,height   The initial size of the visualization
-#' @param elementId   Identifier for the HTML element into which the visualization will be added.
-#' @return A html widget containing a Sankey diagram of the data
+#' @param   alt_click_handler   A JavaScript function to be called when the user alt-clicks on a
+#' node in the Sankey chart. Construct this with \code{htmlwidgets::JS()}. The function should
+#' have parameters "event" and "data".
+#' @param   width,height   The initial size of the visualization
+#' @param   elementId   Identifier for the HTML element into which the visualization will be added.
+#' @return   A html widget containing a Sankey diagram of the data
 #' @examples
 #' data(admissions)
 #' sankey(
@@ -24,7 +27,12 @@
 #' )
 #'
 #' @export
-sankey = function(data, steps, width = NULL, height = NULL, elementId = NULL) {
+sankey = function(data,
+                  steps,
+                  alt_click_handler = NULL,
+                  width = NULL,
+                  height = NULL,
+                  elementId = NULL) {
   used_steps = unlist(steps)
   if (!all(used_steps %in% colnames(data)) || any(duplicated(used_steps))) {
     stop("steps should be unique and be a subset of colnames(data)")
@@ -35,6 +43,10 @@ sankey = function(data, steps, width = NULL, height = NULL, elementId = NULL) {
     data = data[used_steps],
     steps = steps
   )
+
+  if (!is.null(alt_click_handler)) {
+    x$altClickHandler = alt_click_handler
+  }
 
   # Ensures that javascript receives a row-oriented view of 'data'
   attr(x, "TOJSON_ARGS") = list(dataframe = "rows") # nolint: object_name_linter.